// src/app/interfaces/register-user.ts
export class RegisterUser {
    key?: string | null;
    uuid?: string;
    fullname?: string;
    firstname?: string;
    avatar?: string;
    lastname?: string;
    email?: string;
    username?: string;
    passwowrd?: string;
    level?: number;
    phone?: string;
    type?: number;
    published?: boolean;
}